Upgrade to Babel 7
