--- .babelrc
@@ -1,33 +1,0 @@
-{
-    "presets": [
-        ["env", {
-            "targets": {
-                "node": "current"
-            },
-            "useBuiltIns": true
-        }],
-        "react"
-    ],
-    "plugins": [
-        "transform-class-properties",
-        "transform-object-rest-spread"
-    ],
-    "env": {
-        "browser": {
-            "presets": [
-                ["env", {
-                    "targets": {
-                        "browsers": ["last 2 versions"]
-                    },
-                    "modules": false,
-                    "useBuiltIns": true
-                }],
-                "react"
-            ],
-            "plugins": [
-                "transform-class-properties",
-                "transform-object-rest-spread"
-            ]
-        }
-    }
-}
