+++ package.json
@@ -10,15 +10,15 @@
     "config": false
   },
   "scripts": {
-    "dev:server": "webpack-dev-server --config ./src/app/webpack.config.babel.js --host 0.0.0.0",
-    "start": "node --require babel-register src/api",
+    "dev:server": "BABEL_ENV=browser webpack-dev-server --config ./src/app/webpack.config.babel.js --host 0.0.0.0",
+    "start": "node --require @babel/register src/api",
     "build": "BABEL_ENV=browser webpack --config=./src/app/webpack.config.babel.js",
     "dev": "pm2-docker process.yml",
-    "development:api": "nodemon --require babel-register --watch src/api src/api",
+    "development:api": "nodemon --require @babel/register --watch src/api src/api",
     "development:app": "BABEL_ENV=browser node_modules/.bin/webpack-dev-server --config=./src/app/webpack.config.babel.js --port=8080 --host=0.0.0.0",
     "postversion": "git push && git push --tags",
     "lint": "./pre-commit",
-    "test:api:e2e": "mocha --require babel-polyfill --require babel-core/register 'src/api/e2e/*.spec.js'",
+    "test:api:e2e": "mocha --require @babel/polyfill --require babel-core/register 'src/api/e2e/*.spec.js'",
     "test:unit": "./node_modules/jest/bin/jest.js",
     "test:unit:watch": "./node_modules/jest/bin/jest.js --watch"
   },
@@ -39,28 +39,24 @@
   },
   "homepage": "https://github.com/Inist-CNRS/lodex#readme",
   "dependencies": {
+    "@babel/cli": "7.1.2",
+    "@babel/plugin-proposal-class-properties": "7.1.0",
+    "@babel/plugin-proposal-object-rest-spread": "7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "7.0.0",
+    "@babel/plugin-transform-runtime": "7.1.0",
+    "@babel/polyfill": "7.0.0",
+    "@babel/preset-env": "7.1.0",
+    "@babel/preset-react": "7.0.0",
+    "@babel/register": "7.0.0",
     "aphrodite": "1.2.5",
     "async-busboy": "0.6.2",
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.0",
-    "babel-eslint": "8.0.2",
-    "babel-loader": "7.1.2",
-    "babel-plugin-add-module-exports": "0.2.1",
-    "babel-plugin-transform-class-properties": "6.24.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "6.26.0",
-    "babel-plugin-transform-object-rest-spread": "6.26.0",
-    "babel-plugin-transform-runtime": "6.23.0",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-env": "1.6.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-2": "6.24.1",
-    "babel-register": "6.26.0",
-    "babel-runtime": "6.26.0",
+    "babel-eslint": "10.0.1",
+    "babel-loader": "8.0.4",
     "classnames": "2.2.5",
     "comma-number": "2.0.0",
     "common-tags": "1.4.0",
-    "connected-react-router": "4.4.1",
     "config": "1.26.2",
+    "connected-react-router": "4.4.1",
     "csv-parse": "1.2.0",
     "d3": "4.12.2",
     "d3-force": "1.1.0",
@@ -77,9 +73,9 @@
     "fetch-with-proxy": "1.1.0",
     "file-loader": "0.11.2",
     "file-saver": "1.3.3",
-    "htmlparser2": "3.9.2",
     "history": "4.7.2",
     "html2canvas": "1.0.0-alpha.9",
+    "htmlparser2": "3.9.2",
     "inist-ark": "2.1.3",
     "inist-roadmap": "1.2.0",
     "jbj": "7.1.4",
@@ -180,6 +176,9 @@
     "xml-writer": "1.7.0"
   },
   "devDependencies": {
+    "@babel/core": "7.1.2",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-jest": "23.6.0",
     "btoa": "1.1.2",
     "chai": "4.1.2",
     "compression-webpack-plugin": "2.0.0",
@@ -206,7 +205,7 @@
     "prettier": "1.8.2",
     "react-addons-test-utils": "15.6.2",
     "sinon": "4.1.2",
-    "webpack": "4.20.2",
+    "webpack": "4.22.0",
     "webpack-cli": "3.1.2",
     "webpack-dev-server": "3.1.9"
   }
