+++ process.yml
@@ -7,4 +7,4 @@ apps:
         - ./src/api
         - config
         - config.json
-    node_args: --harmony-async-await --require babel-register
+    node_args: --harmony-async-await --require @babel/register
